requirements.txt
# MIT Course Planning RAG Assistant — Python Dependencies
# Python >= 3.11 recommended

# LangChain core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Vector store
faiss-cpu>=1.8.0

# OpenAI
openai>=1.40.0

# Document processing
pypdf>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: Streamlit demo
streamlit>=1.38.0

# Optional: evaluation reporting
tabulate>=0.9.0

streamlit
